[Refactor] (test) Move the base class for testing template into one single class '_spec'.
